@@ -120,7 +120,7 @@
     "cross-env": "7.0.3",
     "detect-port": "1.3.0",
     "dotenv": "8.2.0",
-    "electron": "11.5.0",
+    "electron": "13.6.6",
     "electron-builder": "22.14.13",
     "electron-devtools-installer": "git+https://github.com/MarshallOfSound/electron-devtools-installer.git#31bf8632c32dd6a714b00289708c012a06741fd7",
     "electron-fetch": "1.7.3",
@@ -153,7 +153,7 @@
     "url-loader": "4.1.1",
     "webpack": "5.33.2",
     "webpack-bundle-analyzer": "4.4.1",
-    "webpack-cli": "4.9.0",
+    "webpack-cli": "4.10.0",
     "webpack-dev-server": "3.11.2",
     "webpack-merge": "5.7.3",
     "yarn-deduplicate": "3.1.0"
