@@ -1,12 +1,11 @@
+import installExtension, {
+  REDUX_DEVTOOLS,
+  REACT_DEVELOPER_TOOLS,
+} from 'electron-devtools-installer';
 import { isDebug } from '../utils';
 
 export default async () => {
   if (!isDebug()) return;
-  const {
-    default: installExtension,
-    REDUX_DEVTOOLS,
-    REACT_DEVELOPER_TOOLS,
-  } = require('electron-devtools-installer'); // eslint-disable-line global-require
   await installExtension([REDUX_DEVTOOLS, REACT_DEVELOPER_TOOLS], {
     loadExtensionOptions: { allowFileAccess: true },
     forceDownload: false,
